# target/classes/static/openapi.yaml
openapi: 3.1.1
info:
  title: Bank API
  version: 1.0.0
servers:
  - url: https://localhost:8443
paths:
...
  /cards/my:
    get:
      summary: Get user cards
      security:
        - BearerAuth: [ ]
      tags: [Cards (USER)]
      responses:
        '200':
          description: List of user cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                decryptionError:
                  value: { message: "Decryption error" }
                internalError:
                  value: { message: "Internal server error" }
...
  /cards/{id}/balance:
    get:
      summary: Get card balance
      security:
        - BearerAuth: [ ]
      tags: [Cards (USER)]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Card balance
          content:
            application/json:
              schema:
                type: number
                format: double
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'
        '403':
          description: Card does not belong to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { message: "Card does not belong to user" }
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { message: "Card not found" }
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                decryptionError:
                  value: { message: "Decryption error" }
                internalError:
                  value: { message: "Internal server error" }
  /cards/my/search:
    get:
      summary: Search user cards
      security:
        - BearerAuth: [ ]
      tags: [Cards (USER)]
      parameters:
        - name: number
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Page of user cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDtoPage'
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                decryptionError:
                  value: { message: "Decryption error" }
                internalError:
                  value: { message: "Internal server error" }
  /cards/{id}/block:
    post:
      summary: Request to block card
      security:
        - BearerAuth: [ ]
      tags: [Cards (USER)]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Card was blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'
        '403':
          description: Card does not belong to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { message: "Card does not belong to user" }
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { message: "Card not found" }
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                decryptionError:
                  value: { message: "Decryption error" }
                internalError:
                  value: { message: "Internal server error" }
...
  /users:
    get:
      summary: Get all users
      security:
        - BearerAuth: [ ]
      tags: [Users (ADMIN)]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'          
        '403':
          $ref: '#/components/responses/AccessDenied'
        '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
                example: { message: "Internal server error" }
  /users/{id}:
    get:
      summary: Get user by ID
      security:
        - BearerAuth: [ ]
      tags: [Users (ADMIN)]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'        
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          description: User not found
          content:
            application/json:
              example: { message: "User not found" }
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { message: "Internal server error" }

    delete:
      summary: Delete user
      security:
        - BearerAuth: [ ]
      tags: [Users (ADMIN)]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User was deleted
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'        
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          description: User not found
          content:
            application/json:
              example: { message: "User not found" }
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { message: "Internal server error" }
  /users/{id}/role:
    patch:
      summary: Update user role
      security:
        - BearerAuth: [ ]
      tags: [Users (ADMIN)]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: User role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid request
          content:
            application/json:
              examples:
                badFormat:
                  summary: Invalid request format
                  value: { message: "Invalid request format. Please check your data and try again" }
                role:
                  summary: Role is required
                  value: { message: "Role is required" }
        '401':
          $ref: '#/components/responses/InvalidAccessTokenFormat'          
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
            description: User not found
            content:
              application/json:
                example: { message: "User not found" }
        '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
                example: { message: "Internal server error" }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    AccessDenied:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { status: 403, message: "Access Denied" }
    InvalidAccessTokenFormat:
      description: Invalid access token format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { status: 401, message: "Invalid access token format" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
    UpdateRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [USER, ADMIN]
...
    CardDto:
      type: object
      properties:
        id:
          type: integer
        maskedNumber:
          type: string
        ownerName:
          type: string
        validUntil:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
        balance:
          type: number
          format: double
    CardDtoPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
        totalElements:
          type: integer
        size:
          type: integer
        page:
          type: integer
...